+++ CHANGELOG.md
@@ -33,6 +33,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - `client.batch.hisRead(ids, from, to, options)`
   - `client.batch.hisDelete(ids, range, options)`
   - `client.batch.create(entities, options)`
+  - `client.batch.update(entities, options)`
 
 ### CHANGED
 - Client no longer throws a Axios error if a response has been received and response is a Haystack of GraphQL error.